+++ test/authentication/testOauthSnowflakeAuthorizationCode.js
@@ -2,7 +2,7 @@ const assert = require('assert');
 const connParameters = require('./connectionParameters');
 const AuthTest = require('./authTestsBaseClass.js');
 const authUtil = require('../../lib/authentication/authentication_util');
-const AuthenticationTypes = require('../../lib/authentication/authentication_types');
+const { buildCacheKey, CacheTokenTypes } = require('../../lib/authentication/cache_key_builder');
 
 describe('Oauth Snowflake Authorization code tests', function () {
   const provideBrowserCredentialsPath = '/externalbrowser/provideBrowserCredentials.js';
@@ -79,16 +79,22 @@ describe('Oauth Snowflake Authorization code tests', function () {
       ...connParameters.oauthSnowflakeAuthorizationCode,
       clientStoreTemporaryCredential: true,
     };
-    const accessTokenKey = authUtil.buildOauthAccessTokenCacheKey(
-      connectionOption.host,
-      connectionOption.username,
-      AuthenticationTypes.OAUTH_AUTHORIZATION_CODE,
-    );
-    const refreshTokenKey = authUtil.buildOauthRefreshTokenCacheKey(
-      connectionOption.host,
-      connectionOption.username,
-      AuthenticationTypes.OAUTH_AUTHORIZATION_CODE,
-    );
+    const oauthTokenRequestUrl =
+      connectionOption.oauthTokenRequestUrl || connectionOption.accessUrl + '/oauth/token-request';
+    const accessTokenKey = buildCacheKey({
+      tokenType: CacheTokenTypes.OAUTH_ACCESS_TOKEN,
+      idp: oauthTokenRequestUrl,
+      snowflake: connectionOption.host,
+      username: connectionOption.username,
+      role: connectionOption.role || '',
+    });
+    const refreshTokenKey = buildCacheKey({
+      tokenType: CacheTokenTypes.OAUTH_REFRESH_TOKEN,
+      idp: oauthTokenRequestUrl,
+      snowflake: connectionOption.host,
+      username: connectionOption.username,
+      role: connectionOption.role || '',
+    });
     let firstIdToken;
 
     before(async () => {